Refactor to combine all reducers with combineReducers
