+++ client/src/ports/config/NightWatchConfigRepository.test.js
@@ -5,9 +5,7 @@ const ONE_HOUR_IN_MINUTES = 60;
 
 test('getNightWatchConfig should return the night watch config', () => {
   const state = {
-    config: {
-      maxTotalTimeSpent: ONE_HOUR_IN_MINUTES,
-    },
+    maxTotalTimeSpent: ONE_HOUR_IN_MINUTES,
   };
 
   const repository = new NightWatchConfigRepository(state);
@@ -26,7 +24,7 @@ test('save should persist a state with night watch config attributes', () => {
   });
 
   repository.save(domainConfig);
-  expect(repository.state.config).toHaveProperty('maxTotalTimeSpent', ONE_HOUR_IN_MINUTES);
+  expect(repository.state).toHaveProperty('maxTotalTimeSpent', ONE_HOUR_IN_MINUTES);
 });
 
 test('should not mutate state', () => {
