+++ client/src/store/reducers/ui/switchOffLoadingPlan.test.js
@@ -1,11 +1,13 @@
 import switchOffLoadingPlan from './switchOffLoadingPlan';
-import { initialState } from '../../reducer';
+import { initialState } from './reducer';
 import { getWaitingForPlan } from './selectors';
 
 test('should switch off loading plan', () => {
-  const state = initialState();
-  state.ui.waitingForPlan = true;
-
+  const ui = {
+    ...initialState(),
+    waitingForPlan: true,
+  };
+  const state = { ui };
   const stateAfterAction = switchOffLoadingPlan(state);
   expect(getWaitingForPlan(stateAfterAction)).toBe(false);
 });
