+++ client/src/store/reducers/ui/switchOnLoadingPlan.test.js
@@ -1,10 +1,13 @@
 import switchOnLoadingPlan from './switchOnLoadingPlan';
-import { initialState } from '../../reducer';
+import { initialState } from './reducer';
 import { getWaitingForPlan } from './selectors';
 
 test('should switch on loading plan', () => {
-  const state = initialState();
-  state.ui.waitingForPlan = false;
+  const ui = {
+    ...initialState(),
+    waitingForPlan: false,
+  };
+  const state = { ui };
 
   const stateAfterAction = switchOnLoadingPlan(state);
   expect(getWaitingForPlan(stateAfterAction)).toBe(true);
